dockerfile: use explicit FROM base image
Currently microk8s is bundled with Docker 17.03, and the feature where
ARG is allowed before FROM in a Dockerfile only appeared in 17.05. It is
understood that this version is quite old (at least 1.5 year old) but
to reduce confusion, I'm removing those lines.

Additionally, grabs the binaries from the build directory (currently
compiled binaries are created in the project's root.